refactor: progress 바 while 문 삭제 && 반복 함수 모듈화
