--- .env.example
@@ -1,3 +1,0 @@
-PUBLIC_BASE_URL=http://localhost:5173
-PUBLIC_IMG_API_URL=http://localhost:8000
-PUBLIC_FILE_API_URL=http://localhost:8001
