+++ src/components/RemoveBackground/ImgExample.svelte
